updated file management
